Move AnimationSet to its own source file
