+++ NAS2D/Resources/AnimationSet.cpp
@@ -1,0 +1,32 @@
+
+#include "AnimationSet.h"
+
+#include "../ContainerUtils.h"
+
+
+using namespace NAS2D;
+
+
+AnimationSet::AnimationSet(std::string fileName, std::map<std::string, Image> imageSheets, std::map<std::string, std::vector<AnimationSet::Frame>> actions) :
+	mFileName{std::move(fileName)},
+	mImageSheets{std::move(imageSheets)},
+	mActions{std::move(actions)}
+{
+}
+
+
+std::vector<std::string> AnimationSet::actionNames() const
+{
+	return getKeys(mActions);
+}
+
+
+const std::vector<AnimationSet::Frame>& AnimationSet::frames(const std::string& actionName) const
+{
+	if (mActions.find(actionName) == mActions.end())
+	{
+		throw std::runtime_error("Sprite::play called on undefined action: " + actionName + "  (" + mFileName + ")");
+	}
+
+	return mActions.at(actionName);
+}
