+++ NAS2D/Resources/Sprite.cpp
@@ -12,7 +12,6 @@
 #include "../Utility.h"
 #include "../Filesystem.h"
 #include "../StringUtils.h"
-#include "../ContainerUtils.h"
 #include "../Version.h"
 #include "../Xml/Xml.h"
 #include "../Renderer/Renderer.h"
@@ -52,31 +51,6 @@ namespace {
 		}
 		return iter->second;
 	}
-}
-
-
-AnimationSet::AnimationSet(std::string fileName, std::map<std::string, Image> imageSheets, std::map<std::string, std::vector<AnimationSet::Frame>> actions) :
-	mFileName{std::move(fileName)},
-	mImageSheets{std::move(imageSheets)},
-	mActions{std::move(actions)}
-{
-}
-
-
-std::vector<std::string> AnimationSet::actionNames() const
-{
-	return getKeys(mActions);
-}
-
-
-const std::vector<AnimationSet::Frame>& AnimationSet::frames(const std::string& actionName) const
-{
-	if (mActions.find(actionName) == mActions.end())
-	{
-		throw std::runtime_error("Sprite::play called on undefined action: " + actionName + "  (" + mFileName + ")");
-	}
-
-	return mActions.at(actionName);
 }
