+++ NAS2D/Resources/Sprite.h
@@ -10,44 +10,18 @@
 
 #pragma once
 
-#include "Image.h"
+#include "AnimationSet.h"
 #include "../Signal.h"
 #include "../Timer.h"
 #include "../Renderer/Color.h"
-#include "../Renderer/Rectangle.h"
 
-#include <map>
-#include <string_view>
-#include <string>
 #include <vector>
+#include <string>
 
 
 namespace NAS2D {
 
 constexpr std::string_view SPRITE_VERSION{"0.99"};
-
-
-class AnimationSet
-{
-public:
-	struct Frame
-	{
-		const Image& image;
-		Rectangle<int> bounds;
-		Vector<int> anchorOffset;
-		unsigned int frameDelay;
-	};
-
-	AnimationSet(std::string fileName, std::map<std::string, Image> imageSheets, std::map<std::string, std::vector<Frame>> actions);
-
-	std::vector<std::string> actionNames() const;
-	const std::vector<Frame>& frames(const std::string& actionName) const;
-
-private:
-	std::string mFileName;
-	std::map<std::string, Image> mImageSheets;
-	std::map<std::string, std::vector<Frame>> mActions;
-};
 
 
 /**
